Made `build.sh` more resiliant, letting it resume after error.

+++ bin/build.sh
@@ -219,14 +219,18 @@ then
     ;;
 
   linux-ubuntu-14*)
-    sudo add-apt-repository "deb http://llvm.org/apt/trusty/ llvm-toolchain-trusty-3.5 main"
-    ${WGET} -O - http://llvm.org/apt/llvm-snapshot.gpg.key | sudo apt-key add -
-    sudo apt-get update
-    sudo apt-get install -y ${DEPENDENCIES}
-    sudo update-alternatives --install /usr/bin/clang clang /usr/bin/clang-3.5 20
-    sudo update-alternatives --install /usr/bin/clang++ clang++ /usr/bin/clang++-3.5 20
-    sudo update-alternatives --install /usr/bin/llvm-config llvm-config /usr/bin/llvm-config-3.5 20
-    sudo update-alternatives --install /usr/bin/llvm-link llvm-link /usr/bin/llvm-link-3.5 20
+    if ! which clang-3.5 > /dev/null; then
+      sudo add-apt-repository "deb http://llvm.org/apt/trusty/ llvm-toolchain-trusty-3.5 main"
+      ${WGET} -O - http://llvm.org/apt/llvm-snapshot.gpg.key | sudo apt-key add -
+      sudo apt-get update
+      sudo apt-get install -y ${DEPENDENCIES}
+      sudo update-alternatives --install /usr/bin/clang clang /usr/bin/clang-3.5 20
+      sudo update-alternatives --install /usr/bin/clang++ clang++ /usr/bin/clang++-3.5 20
+      sudo update-alternatives --install /usr/bin/llvm-config llvm-config /usr/bin/llvm-config-3.5 20
+      sudo update-alternatives --install /usr/bin/llvm-link llvm-link /usr/bin/llvm-link-3.5 20
+    else
+      puts "Required packages already installed."
+    fi
     ;;
 
   linux-ubuntu-12*)
@@ -310,13 +314,16 @@ fi
 if [ ${BUILD_Z3} -eq 1 ]
 then
   puts "Installing Z3"
+  if [ ! -f ${Z3_DIR}/bin/z3 ]; then
+    ${WGET} ${Z3_DOWNLOAD_LINK} -O z3-downloaded.zip
+    unzip -o z3-downloaded.zip -d z3-extracted
+    mv z3-extracted/z3-* ${Z3_DIR}
+    rm -rf z3-downloaded.zip z3-extracted
 
-  ${WGET} ${Z3_DOWNLOAD_LINK} -O z3-downloaded.zip
-  unzip -o z3-downloaded.zip -d z3-extracted
-  mv z3-extracted/z3-* ${Z3_DIR}
-  rm -rf z3-downloaded.zip z3-extracted
-
-  puts "Installed Z3"
+    puts "Installed Z3"
+  else
+    puts "Z3 already installed"
+  fi
 fi
 
 
@@ -324,18 +331,22 @@ if [ ${BUILD_BOOGIE} -eq 1 ]
 then
   puts "Building Boogie"
 
-  git clone https://github.com/boogie-org/boogie.git ${BOOGIE_DIR}
-  cd ${BOOGIE_DIR}
-  git checkout ${BOOGIE_COMMIT}
-  cd ${BOOGIE_DIR}/Source
-  mozroots --import --sync
-  ${WGET} https://nuget.org/nuget.exe
-  mono ./nuget.exe restore Boogie.sln
-  rm -rf /tmp/nuget/
-  xbuild Boogie.sln /p:Configuration=Release
-  ln -s ${Z3_DIR}/bin/z3 ${BOOGIE_DIR}/Binaries/z3.exe
+  if [ ! -f ${BOOGIE_DIR}/Binaries/Boogie.exe ]; then
+    git clone https://github.com/boogie-org/boogie.git ${BOOGIE_DIR}
+    cd ${BOOGIE_DIR}
+    git checkout ${BOOGIE_COMMIT}
+    cd ${BOOGIE_DIR}/Source
+    mozroots --import --sync
+    ${WGET} https://nuget.org/nuget.exe
+    mono ./nuget.exe restore Boogie.sln
+    rm -rf /tmp/nuget/
+    xbuild Boogie.sln /p:Configuration=Release
+    ln -s ${Z3_DIR}/bin/z3 ${BOOGIE_DIR}/Binaries/z3.exe
 
-  puts "Built Boogie"
+    puts "Built Boogie"
+  else
+    puts "Boogie already installed"
+  fi
 fi
 
 
@@ -343,47 +354,56 @@ if [ ${BUILD_CORRAL} -eq 1 ]
 then
   puts "Building Corral"
 
-  cd ${ROOT}
-  ${WGET} ${CORRAL_DOWNLOAD_LINK} -O corral-downloaded.zip
-  unzip -o corral-downloaded.zip -d ${CORRAL_DIR}
-  rm -f corral-downloaded.zip
-  cd ${CORRAL_DIR}
-#  git clone https://git01.codeplex.com/corral ${CORRAL_DIR}
-#  cd ${CORRAL_DIR}
-#  git checkout ${CORRAL_COMMIT}
-  cp ${BOOGIE_DIR}/Binaries/*.{dll,exe} references
-  xbuild cba.sln /p:Configuration=Release
-  ln -s ${Z3_DIR}/bin/z3 ${CORRAL_DIR}/bin/Release/z3.exe
+  if [ ! -f ${CORRAL_DIR}/bin/Release/corral.exe ]; then
+    cd ${ROOT}
+    ${WGET} ${CORRAL_DOWNLOAD_LINK} -O corral-downloaded.zip
+    unzip -o corral-downloaded.zip -d ${CORRAL_DIR}
+    rm -f corral-downloaded.zip
+    cd ${CORRAL_DIR}
+#    git clone https://git01.codeplex.com/corral ${CORRAL_DIR}
+#    cd ${CORRAL_DIR}
+#    git checkout ${CORRAL_COMMIT}
+    cp ${BOOGIE_DIR}/Binaries/*.{dll,exe} references
+    xbuild cba.sln /p:Configuration=Release
+    ln -s ${Z3_DIR}/bin/z3 ${CORRAL_DIR}/bin/Release/z3.exe
 
-  puts "Built Corral"
+    puts "Built Corral"
+  else
+    puts "Corral already installed"
+  fi
 fi
 
 
 if [ ${BUILD_SMACK} -eq 1 ]
 then
   puts "Building SMACK"
+  
+  if ! which smack > /dev/null; then
+    mkdir -p ${SMACK_DIR}/build
+    cd ${SMACK_DIR}/build
+    cmake ${CMAKE_INSTALL_PREFIX} -DCMAKE_C_COMPILER=clang -DCMAKE_CXX_COMPILER=clang++ -DCMAKE_BUILD_TYPE=Debug ..
+    make
+    sudo make install
 
-  mkdir -p ${SMACK_DIR}/build
-  cd ${SMACK_DIR}/build
-  cmake ${CMAKE_INSTALL_PREFIX} -DCMAKE_C_COMPILER=clang -DCMAKE_CXX_COMPILER=clang++ -DCMAKE_BUILD_TYPE=Debug ..
-  make
-  sudo make install
+    puts "Configuring shell environment"
+    echo export BOOGIE=\"mono ${BOOGIE_DIR}/Binaries/Boogie.exe\" >> ${SMACKENV}
+    echo export CORRAL=\"mono ${CORRAL_DIR}/bin/Release/corral.exe\" >> ${SMACKENV}
+    source ${SMACKENV}
+    puts "The required environment variables have been set in ${SMACKENV}"
+    puts "You should source ${SMACKENV} in your .bashrc"
 
-  puts "Configuring shell environment"
-  echo export BOOGIE=\"mono ${BOOGIE_DIR}/Binaries/Boogie.exe\" >> ${SMACKENV}
-  echo export CORRAL=\"mono ${CORRAL_DIR}/bin/Release/corral.exe\" >> ${SMACKENV}
-  source ${SMACKENV}
-  puts "The required environment variables have been set in ${SMACKENV}"
-  puts "You should source ${SMACKENV} in your .bashrc"
-
-  puts "Built SMACK"
+    puts "Built SMACK"
+  else
+    puts "SMACK already installed"
+  fi
 fi
 
 
 if [ ${TEST_SMACK} -eq 1 ]
 then
   puts "Running SMACK regression tests"
-
+  
+  source ${SMACKENV} # Make sure we can run the tests
   cd ${SMACK_DIR}/test
   ./regtest.py
   res=$?
